requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
mariadb==1.0.7
MarkupSafe==2.0.1
PyYAML==6.0
SQLAlchemy==1.4.25
Werkzeug==2.0.1
